+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 20.10.7
+
+### Bugs
+
+*logs*
+
+There was a possible deadlock in logging::broker due to creation of logs when
+sending logs, it was then possible to lock a mutex twice.
+
 ## 20.10.6
 
 ### Bugs